extract loaders and adjust middleware
